Gislaine quinta

+++ cookenu1/src/index.ts
@@ -155,6 +155,13 @@ app.post("/user/follow", async (req: Request, res: Response) => {
     const id_follower = authenticationData.id;
     const id_followed = req.body.userToFollowId
 
+    const followId = new FollowDatabase();
+    const idDb = await followId.isValidId(id_followed);
+
+    if (idDb.quantity === 0) {
+      throw new Error("Insira um id válido");
+    }
+    
     const followDb = new FollowDatabase();   
     await followDb.createFollow(id_followed, id_follower);   
 
@@ -168,7 +175,7 @@ app.post("/user/follow", async (req: Request, res: Response) => {
   }
 });
 
-app.post("/user/unfollow", async (req: Request, res: Response) => {
+app.delete("/user/unfollow", async (req: Request, res: Response) => {
   try {
     if (!req.body.userToUnfollowId || req.body.userToUnfollowId === " ") {
       throw new Error("Insira um id");
@@ -181,6 +188,12 @@ app.post("/user/unfollow", async (req: Request, res: Response) => {
     const id_follower = authenticationData.id;
     const id_followed = req.body.userToUnfollowId
 
+    const followId = new FollowDatabase();
+    const idDb = await followId.isValidId(id_followed);
+    
+    if (idDb.quantity === 0) {
+      throw new Error("Insira um id válido");
+    }
     const followDb = new FollowDatabase();   
     await followDb.deleteFollow(id_followed, id_follower);
